Remove editing of a song's parent playlist
The UI currently provides no affordances for the user to edit the
playlist that a song belongs to. The playlistName property of the song
will now be omitted from the request body. The backend will not
check if a song's new playlist is equal to the old one, speeding up the
edit process.

The editing of a song's playlist was also introducing a bug that would
remove the song from the playlist's "song" array. In the future when
this feature is added again, after the UI allows the user to do so, the
bug will have to be addressed.